+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.5.0 (in development)
+-----------------------
+- `run` and `run-pr`: Shebang handling for the `--script` option is now done by
+  `forall` directly rather than by invoking `perl`
+
 v0.4.0 (2025-11-01)
 -------------------
 - Renamed `--skip` to `--exclude`